Rename `constants.new`
